requirements/tests.txt
coverage==3.7.1
mock==0.8.0
django-celery==3.1.17
# python-statsd==1.5.4
redis==2.6.2
